+++ src/main/java/com/nha/java/learning/phoneshop/repository/BrandRepository.java
@@ -9,10 +9,11 @@ import org.springframework.stereotype.Repository;
 import com.nha.java.learning.phoneshop.entity.Brand;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>, 
+public interface BrandRepository extends JpaRepository<Brand, Long>, 
 				JpaSpecificationExecutor<Brand>{
 	
 	List<Brand> findByNameIgnoreCase(String name);
 	List<Brand> findByNameLike(String name);
 	List<Brand> findByNameContaining(String name);
+	
 }
